Substitue deprecate types with new ones
