+++ client/keys/delete_test.go
@@ -1,8 +1,6 @@
 package keys
 
 import (
-	"bufio"
-	"strings"
 	"testing"
 
 	"github.com/spf13/viper"
@@ -20,18 +18,12 @@ func Test_runDeleteCmd(t *testing.T) {
 
 	yesF, _ := deleteKeyCommand.Flags().GetBool(flagYes)
 	forceF, _ := deleteKeyCommand.Flags().GetBool(flagForce)
+
 	require.False(t, yesF)
 	require.False(t, forceF)
 
 	fakeKeyName1 := "runDeleteCmd_Key1"
 	fakeKeyName2 := "runDeleteCmd_Key2"
-	kb, err := keyring.New(sdk.KeyringServiceName(), viper.GetString(flags.FlagKeyringBackend), viper.GetString(flags.FlagHome), mockIn)
-	require.NoError(t, err)
-	t.Cleanup(func() {
-		kb.Delete("runDeleteCmd_Key1") // nolint:errcheck
-		kb.Delete("runDeleteCmd_Key2") // nolint:errcheck
-
-	})
 	// Now add a temporary keybase
 	kbHome, cleanUp := tests.NewTestCaseDir(t)
 	t.Cleanup(cleanUp)
@@ -39,12 +31,12 @@ func Test_runDeleteCmd(t *testing.T) {
 
 	// Now
 	path := sdk.GetConfig().GetFullFundraiserPath()
-	kb, err = keyring.New(sdk.KeyringServiceName(), viper.GetString(flags.FlagKeyringBackend), kbHome, mockIn)
+	backend := viper.GetString(flags.FlagKeyringBackend)
+	kb, err := keyring.New(sdk.KeyringServiceName(), backend, kbHome, mockIn)
 	require.NoError(t, err)
 	_, err = kb.NewAccount(fakeKeyName1, tests.TestMnemonic, "", path, keyring.AltSecp256k1)
 	require.NoError(t, err)
-
-	_, err = kb.NewAccount(fakeKeyName2, tests.TestMnemonic, "", path, keyring.AltSecp256k1)
+	_, _, err = kb.NewMnemonic(fakeKeyName2, keyring.English, keyring.AltSecp256k1)
 	require.NoError(t, err)
 
 	err = runDeleteCmd(deleteKeyCommand, []string{"blah"})
@@ -56,17 +48,15 @@ func Test_runDeleteCmd(t *testing.T) {
 	require.Error(t, err)
 	require.Equal(t, "EOF", err.Error())
 
-	{
-		_, err = kb.Key(fakeKeyName1)
-		require.NoError(t, err)
+	_, err = kb.Key(fakeKeyName1)
+	require.NoError(t, err)
 
-		// Now there is a confirmation
-		viper.Set(flagYes, true)
-		require.NoError(t, runDeleteCmd(deleteKeyCommand, []string{fakeKeyName1}))
+	// Now there is a confirmation
+	viper.Set(flagYes, true)
+	require.NoError(t, runDeleteCmd(deleteKeyCommand, []string{fakeKeyName1}))
 
-		_, err = kb.Key(fakeKeyName1)
-		require.Error(t, err) // Key1 is gone
-	}
+	_, err = kb.Key(fakeKeyName1)
+	require.Error(t, err) // Key1 is gone
 
 	viper.Set(flagYes, true)
 	_, err = kb.Key(fakeKeyName2)
@@ -75,34 +65,4 @@ func Test_runDeleteCmd(t *testing.T) {
 	require.NoError(t, err)
 	_, err = kb.Key(fakeKeyName2)
 	require.Error(t, err) // Key2 is gone
-}
-
-func Test_confirmDeletion(t *testing.T) {
-	type args struct {
-		buf *bufio.Reader
-	}
-
-	answerYes := bufio.NewReader(strings.NewReader("y\n"))
-	answerYes2 := bufio.NewReader(strings.NewReader("Y\n"))
-	answerNo := bufio.NewReader(strings.NewReader("n\n"))
-	answerInvalid := bufio.NewReader(strings.NewReader("245\n"))
-
-	tests := []struct {
-		name    string
-		args    args
-		wantErr bool
-	}{
-		{"Y", args{answerYes}, false},
-		{"y", args{answerYes2}, false},
-		{"N", args{answerNo}, true},
-		{"BAD", args{answerInvalid}, true},
-	}
-	for _, tt := range tests {
-		tt := tt
-		t.Run(tt.name, func(t *testing.T) {
-			if err := confirmDeletion(tt.args.buf); (err != nil) != tt.wantErr {
-				t.Errorf("confirmDeletion() error = %v, wantErr %v", err, tt.wantErr)
-			}
-		})
-	}
 }
