+++ crypto/keyring/keyring.go
@@ -2,6 +2,7 @@ package keyring
 
 import (
 	"bufio"
+	"encoding/hex"
 	"fmt"
 	"io"
 	"io/ioutil"
@@ -11,15 +12,16 @@ import (
 	"strings"
 
 	"github.com/99designs/keyring"
+	"github.com/cosmos/go-bip39"
 	"github.com/pkg/errors"
-
 	"github.com/tendermint/crypto/bcrypt"
 	tmcrypto "github.com/tendermint/tendermint/crypto"
 	cryptoAmino "github.com/tendermint/tendermint/crypto/encoding/amino"
 
 	"github.com/cosmos/cosmos-sdk/client/input"
 	"github.com/cosmos/cosmos-sdk/crypto"
-	"github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/crypto/keys/hd"
+	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 )
 
@@ -38,6 +40,11 @@ const (
 	passKeyringPrefix     = keyringDirNameFmt
 )
 
+var (
+	_                          Keyring = &keystore{}
+	maxPassphraseEntryAttempts         = 3
+)
+
 // Keyring exposes operations over a backend supported by github.com/99designs/keyring.
 type Keyring interface {
 	// List all keys.
@@ -45,23 +52,23 @@ type Keyring interface {
 
 	// Key and KeyByAddress return keys by uid and address respectively.
 	Key(uid string) (Info, error)
-	KeyByAddress(address types.Address) (Info, error)
+	KeyByAddress(address sdk.Address) (Info, error)
 
 	// Delete and DeleteByAddress remove keys from the keyring.
 	Delete(uid string) error
-	DeleteByAddress(address types.Address) error
+	DeleteByAddress(address sdk.Address) error
 
 	// NewMnemonic generates a new mnemonic, derives a hierarchical deterministic
 	// key from that, and persists it to the storage. Returns the generated mnemonic and the key
 	// Info. It returns an error if it fails to generate a key for the given algo type, or if
 	// another key is already stored under the same name.
-	NewMnemonic(uid string, language Language, algo AltSigningAlgo) (Info, string, error)
+	NewMnemonic(uid string, language Language, algo SigningAlgo) (Info, string, error)
 
 	// NewAccount converts a mnemonic to a private key and BIP-39 HD Path and persists it.
-	NewAccount(uid, mnemonic, bip39Passwd, hdPath string, algo AltSigningAlgo) (Info, error)
+	NewAccount(uid, mnemonic, bip39Passwd, hdPath string, algo SigningAlgo) (Info, error)
 
 	// SaveLedgerKey retrieves a public key reference from a Ledger device and persists it.
-	SaveLedgerKey(uid string, algo AltSigningAlgo, hrp string, account, index uint32) (Info, error)
+	SaveLedgerKey(uid string, algo SigningAlgo, hrp string, account, index uint32) (Info, error)
 
 	// SavePubKey stores a public key and returns the persisted Info structure.
 	SavePubKey(uid string, pubkey tmcrypto.PubKey, algo pubKeyType) (Info, error)
@@ -81,7 +88,7 @@ type Signer interface {
 	Sign(uid string, msg []byte) ([]byte, tmcrypto.PubKey, error)
 
 	// SignByAddress sign byte messages with a user key providing the address.
-	SignByAddress(address types.Address, msg []byte) ([]byte, tmcrypto.PubKey, error)
+	SignByAddress(address sdk.Address, msg []byte) ([]byte, tmcrypto.PubKey, error)
 }
 
 // Importer is implemented by key stores that support import of public and private keys.
@@ -94,44 +101,35 @@ type Importer interface {
 type Exporter interface {
 	// Export public key
 	ExportPubKeyArmor(uid string) (string, error)
-	ExportPubKeyArmorByAddress(address types.Address) (string, error)
+	ExportPubKeyArmorByAddress(address sdk.Address) (string, error)
 	// ExportPrivKey returns a private key in ASCII armored format.
 	// It returns an error if the key does not exist or a wrong encryption passphrase is supplied.
 	ExportPrivKeyArmor(uid, encryptPassphrase string) (armor string, err error)
-	ExportPrivKeyArmorByAddress(address types.Address, encryptPassphrase string) (armor string, err error)
+	ExportPrivKeyArmorByAddress(address sdk.Address, encryptPassphrase string) (armor string, err error)
 }
 
 type AltKeyringOption func(options *altKrOptions)
 
-// keyringKeybase implements the Keybase interface by using the Keyring library
-// for account key persistence.
-type keyringKeybase struct {
-	base baseKeybase
-	db   keyring.Keyring
+// NewInMemory creates a transient keyring useful for testing
+// purposes and on-the-fly key generation.
+// Keybase options can be applied when generating this new Keybase.
+func NewInMemory(opts ...AltKeyringOption) Keyring {
+	return newKeystore(keyring.NewArrayKeyring(nil), opts...)
 }
 
-var maxPassphraseEntryAttempts = 3
-
-func newKeyringKeybase(db keyring.Keyring, opts ...KeybaseOption) Keybase {
-	return keyringKeybase{
-		db:   db,
-		base: newBaseKeybase(opts...),
-	}
-}
-
-// NewKeyring creates a new instance of a keyring. Keybase
-// options can be applied when generating this new Keybase.
-// Available backends are "os", "file", "kwallet", "pass", "test".
-func NewKeyring(
-	appName, backend, rootDir string, userInput io.Reader, opts ...KeybaseOption,
-) (Keybase, error) {
+// NewKeyring creates a new instance of a keyring.
+// Keyring ptions can be applied when generating the new instance.
+// Available backends are "os", "file", "kwallet", "memory", "pass", "test".
+func New(
+	appName, backend, rootDir string, userInput io.Reader, opts ...AltKeyringOption,
+) (Keyring, error) {
 
 	var db keyring.Keyring
 	var err error
 
 	switch backend {
 	case BackendMemory:
-		panic("not implemented")
+		return NewInMemory(opts...), err
 	case BackendTest:
 		db, err = keyring.Open(lkbToKeyringConfig(appName, rootDir, nil, true))
 	case BackendFile:
@@ -145,168 +143,72 @@ func NewKeyring(
 	default:
 		return nil, fmt.Errorf("unknown keyring backend %v", backend)
 	}
+
 	if err != nil {
 		return nil, err
 	}
 
-	return newKeyringKeybase(db, opts...), nil
+	return newKeystore(db, opts...), nil
 }
 
-// NewInMemory creates a transient keyring useful for testing
-// purposes and on-the-fly key generation.
-// Keybase options can be applied when generating this new Keybase.
-func NewInMemory(opts ...AltKeyringOption) Keyring {
-	return newKeystore(keyring.NewArrayKeyring(nil), opts...)
+type keystore struct {
+	db      keyring.Keyring
+	options altKrOptions
 }
 
-// CreateMnemonic generates a new key and persists it to storage, encrypted
-// using the provided password. It returns the generated mnemonic and the key Info.
-// An error is returned if it fails to generate a key for the given algo type,
-// or if another key is already stored under the same name.
-func (kb keyringKeybase) CreateMnemonic(
-	name string, language Language, passwd string, algo pubKeyType,
-) (info Info, mnemonic string, err error) {
+func newKeystore(kr keyring.Keyring, opts ...AltKeyringOption) keystore {
+	// Default options for keybase
+	options := altKrOptions{
+		supportedAlgos:       AltSigningAlgoList{AltSecp256k1},
+		supportedAlgosLedger: AltSigningAlgoList{AltSecp256k1},
+	}
 
-	return kb.base.CreateMnemonic(kb, name, language, passwd, algo)
+	for _, optionFn := range opts {
+		optionFn(&options)
+	}
+
+	return keystore{kr, options}
 }
 
-// CreateAccount converts a mnemonic to a private key and persists it, encrypted
-// with the given password.
-func (kb keyringKeybase) CreateAccount(
-	name, mnemonic, bip39Passwd, encryptPasswd, hdPath string, algo pubKeyType,
-) (Info, error) {
-
-	return kb.base.CreateAccount(kb, name, mnemonic, bip39Passwd, encryptPasswd, hdPath, algo)
-}
-
-// CreateLedger creates a new locally-stored reference to a Ledger keypair.
-// It returns the created key info and an error if the Ledger could not be queried.
-func (kb keyringKeybase) CreateLedger(
-	name string, algo pubKeyType, hrp string, account, index uint32,
-) (Info, error) {
-
-	return kb.base.CreateLedger(kb, name, algo, hrp, account, index)
-}
-
-// CreateOffline creates a new reference to an offline keypair. It returns the
-// created key info.
-func (kb keyringKeybase) CreateOffline(name string, pub tmcrypto.PubKey, algo pubKeyType) (Info, error) {
-	return kb.base.writeOfflineKey(kb, name, pub, algo), nil
-}
-
-// CreateMulti creates a new reference to a multisig (offline) keypair. It
-// returns the created key Info object.
-func (kb keyringKeybase) CreateMulti(name string, pub tmcrypto.PubKey) (Info, error) {
-	return kb.base.writeMultisigKey(kb, name, pub), nil
-}
-
-// List returns the keys from storage in alphabetical order.
-func (kb keyringKeybase) List() ([]Info, error) {
-	var res []Info
-	keys, err := kb.db.Keys()
+func (ks keystore) ExportPubKeyArmor(uid string) (string, error) {
+	bz, err := ks.Key(uid)
 	if err != nil {
-		return nil, err
+		return "", err
 	}
 
-	sort.Strings(keys)
-
-	for _, key := range keys {
-		if strings.HasSuffix(key, infoSuffix) {
-			rawInfo, err := kb.db.Get(key)
-			if err != nil {
-				return nil, err
-			}
-
-			if len(rawInfo.Data) == 0 {
-				return nil, sdkerrors.Wrap(sdkerrors.ErrKeyNotFound, key)
-			}
-
-			info, err := unmarshalInfo(rawInfo.Data)
-			if err != nil {
-				return nil, err
-			}
-
-			res = append(res, info)
-		}
+	if bz == nil {
+		return "", fmt.Errorf("no key to export with name: %s", uid)
 	}
 
-	return res, nil
+	return crypto.ArmorPubKeyBytes(bz.GetPubKey().Bytes(), string(bz.GetAlgo())), nil
 }
 
-// Get returns the public information about one key.
-func (kb keyringKeybase) Get(name string) (Info, error) {
-	key := infoKey(name)
-
-	bs, err := kb.db.Get(string(key))
+func (ks keystore) ExportPubKeyArmorByAddress(address sdk.Address) (string, error) {
+	info, err := ks.KeyByAddress(address)
 	if err != nil {
-		return nil, err
+		return "", err
 	}
 
-	if len(bs.Data) == 0 {
-		return nil, sdkerrors.Wrap(sdkerrors.ErrKeyNotFound, name)
-	}
-
-	return unmarshalInfo(bs.Data)
+	return ks.ExportPubKeyArmor(info.GetName())
 }
 
-// GetByAddress fetches a key by address and returns its public information.
-func (kb keyringKeybase) GetByAddress(address types.AccAddress) (Info, error) {
-	ik, err := kb.db.Get(string(addrHexKey(address)))
+func (ks keystore) ExportPrivKeyArmor(uid, encryptPassphrase string) (armor string, err error) {
+	priv, err := ks.ExportPrivateKeyObject(uid)
 	if err != nil {
-		return nil, err
+		return "", err
 	}
 
-	if len(ik.Data) == 0 {
-		return nil, fmt.Errorf("key with address %s not found", address)
-	}
-
-	bs, err := kb.db.Get(string(ik.Data))
+	info, err := ks.Key(uid)
 	if err != nil {
-		return nil, err
+		return "", err
 	}
 
-	return unmarshalInfo(bs.Data)
-}
-
-// Sign signs an arbitrary set of bytes with the named key. It returns an error
-// if the key doesn't exist or the decryption fails.
-func (kb keyringKeybase) Sign(name, passphrase string, msg []byte) (sig []byte, pub tmcrypto.PubKey, err error) {
-	info, err := kb.Get(name)
-	if err != nil {
-		return
-	}
-
-	var priv tmcrypto.PrivKey
-
-	switch i := info.(type) {
-	case localInfo:
-		if i.PrivKeyArmor == "" {
-			return nil, nil, fmt.Errorf("private key not available")
-		}
-
-		priv, err = cryptoAmino.PrivKeyFromBytes([]byte(i.PrivKeyArmor))
-		if err != nil {
-			return nil, nil, err
-		}
-
-	case ledgerInfo:
-		return SignWithLedger(info, msg)
-
-	case offlineInfo, multiInfo:
-		return nil, info.GetPubKey(), errors.New("cannot sign with offline keys")
-	}
-
-	sig, err = priv.Sign(msg)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	return sig, priv.PubKey(), nil
+	return crypto.EncryptArmorPrivKey(priv, encryptPassphrase, string(info.GetAlgo())), nil
 }
 
 // ExportPrivateKeyObject exports an armored private key object.
-func (kb keyringKeybase) ExportPrivateKeyObject(name string, passphrase string) (tmcrypto.PrivKey, error) {
-	info, err := kb.Get(name)
+func (ks keystore) ExportPrivateKeyObject(uid string) (tmcrypto.PrivKey, error) {
+	info, err := ks.Key(uid)
 	if err != nil {
 		return nil, err
 	}
@@ -332,92 +234,18 @@ func (kb keyringKeybase) ExportPrivateKeyObject(name string, passphrase string) 
 	return priv, nil
 }
 
-// Export exports armored private key to the caller.
-func (kb keyringKeybase) Export(name string) (armor string, err error) {
-	bz, err := kb.db.Get(string(infoKey(name)))
+func (ks keystore) ExportPrivKeyArmorByAddress(address sdk.Address, encryptPassphrase string) (armor string, err error) {
+	byAddress, err := ks.KeyByAddress(address)
 	if err != nil {
 		return "", err
 	}
 
-	if bz.Data == nil {
-		return "", fmt.Errorf("no key to export with name: %s", name)
-	}
-
-	return crypto.ArmorInfoBytes(bz.Data), nil
+	return ks.ExportPrivKeyArmor(byAddress.GetName(), encryptPassphrase)
 }
 
-// ExportPubKey returns public keys in ASCII armored format. It retrieves an Info
-// object by its name and return the public key in a portable format.
-func (kb keyringKeybase) ExportPubKey(name string) (armor string, err error) {
-	bz, err := kb.Get(name)
-	if err != nil {
-		return "", err
-	}
-
-	if bz == nil {
-		return "", fmt.Errorf("no key to export with name: %s", name)
-	}
-
-	return crypto.ArmorPubKeyBytes(bz.GetPubKey().Bytes(), string(bz.GetAlgo())), nil
-}
-
-// Import imports armored private key.
-func (kb keyringKeybase) Import(name string, armor string) error {
-	bz, _ := kb.Get(name)
-
-	if bz != nil {
-		pubkey := bz.GetPubKey()
-
-		if len(pubkey.Bytes()) > 0 {
-			return fmt.Errorf("cannot overwrite data for name: %s", name)
-		}
-	}
-
-	infoBytes, err := crypto.UnarmorInfoBytes(armor)
-	if err != nil {
-		return err
-	}
-
-	info, err := unmarshalInfo(infoBytes)
-	if err != nil {
-		return err
-	}
-
-	kb.writeInfo(name, info)
-
-	err = kb.db.Set(keyring.Item{
-		Key:  string(addrHexKey(info.GetAddress())),
-		Data: infoKey(name),
-	})
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// ExportPrivKey returns a private key in ASCII armored format. An error is returned
-// if the key does not exist or a wrong encryption passphrase is supplied.
-func (kb keyringKeybase) ExportPrivKey(name, decryptPassphrase, encryptPassphrase string) (armor string, err error) {
-	priv, err := kb.ExportPrivateKeyObject(name, decryptPassphrase)
-	if err != nil {
-		return "", err
-	}
-
-	info, err := kb.Get(name)
-	if err != nil {
-		return "", err
-	}
-
-	return crypto.EncryptArmorPrivKey(priv, encryptPassphrase, string(info.GetAlgo())), nil
-}
-
-// ImportPrivKey imports a private key in ASCII armor format. An error is returned
-// if a key with the same name exists or a wrong encryption passphrase is
-// supplied.
-func (kb keyringKeybase) ImportPrivKey(name, armor, passphrase string) error {
-	if kb.HasKey(name) {
-		return fmt.Errorf("cannot overwrite key: %s", name)
+func (ks keystore) ImportPrivKey(uid, armor, passphrase string) error {
+	if _, err := ks.Key(uid); err == nil {
+		return fmt.Errorf("cannot overwrite key: %s", uid)
 	}
 
 	privKey, algo, err := crypto.UnarmorDecryptPrivKey(armor, passphrase)
@@ -425,28 +253,17 @@ func (kb keyringKeybase) ImportPrivKey(name, armor, passphrase string) error {
 		return errors.Wrap(err, "failed to decrypt private key")
 	}
 
-	// NOTE: The keyring keystore has no need for a passphrase.
-	kb.writeLocalKey(name, privKey, pubKeyType(algo))
+	_, err = ks.writeLocalKey(uid, privKey, pubKeyType(algo))
+	if err != nil {
+		return err
+	}
+
 	return nil
 }
 
-// HasKey returns whether the key exists in the keyring.
-func (kb keyringKeybase) HasKey(name string) bool {
-	bz, _ := kb.Get(name)
-	return bz != nil
-}
-
-// ImportPubKey imports an ASCII-armored public key. It will store a new Info
-// object holding a public key only, i.e. it will not be possible to sign with
-// it as it lacks the secret key.
-func (kb keyringKeybase) ImportPubKey(name string, armor string) error {
-	bz, _ := kb.Get(name)
-	if bz != nil {
-		pubkey := bz.GetPubKey()
-
-		if len(pubkey.Bytes()) > 0 {
-			return fmt.Errorf("cannot overwrite data for name: %s", name)
-		}
+func (ks keystore) ImportPubKey(uid string, armor string) error {
+	if _, err := ks.Key(uid); err == nil {
+		return fmt.Errorf("cannot overwrite key: %s", uid)
 	}
 
 	pubBytes, algo, err := crypto.UnarmorPubKeyBytes(armor)
@@ -459,27 +276,7 @@ func (kb keyringKeybase) ImportPubKey(name string, armor string) error {
 		return err
 	}
 
-	kb.base.writeOfflineKey(kb, name, pubKey, pubKeyType(algo))
-	return nil
-}
-
-// Delete removes key forever, but we must present the proper passphrase before
-// deleting it (for security). It returns an error if the key doesn't exist or
-// passphrases don't match. The passphrase is ignored when deleting references to
-// offline and Ledger / HW wallet keys.
-func (kb keyringKeybase) Delete(name, _ string, _ bool) error {
-	// verify we have the proper password before deleting
-	info, err := kb.Get(name)
-	if err != nil {
-		return err
-	}
-
-	err = kb.db.Remove(string(addrHexKey(info.GetAddress())))
-	if err != nil {
-		return err
-	}
-
-	err = kb.db.Remove(string(infoKey(name)))
+	_, err = ks.writeOfflineKey(uid, pubKey, pubKeyType(algo))
 	if err != nil {
 		return err
 	}
@@ -487,45 +284,233 @@ func (kb keyringKeybase) Delete(name, _ string, _ bool) error {
 	return nil
 }
 
-// SupportedAlgos returns a list of supported signing algorithms.
-func (kb keyringKeybase) SupportedAlgos() []pubKeyType {
-	return kb.base.SupportedAlgos()
-}
-
-// SupportedAlgosLedger returns a list of supported ledger signing algorithms.
-func (kb keyringKeybase) SupportedAlgosLedger() []pubKeyType {
-	return kb.base.SupportedAlgosLedger()
-}
-
-func (kb keyringKeybase) writeLocalKey(name string, priv tmcrypto.PrivKey, algo pubKeyType) Info {
-	// encrypt private key using keyring
-	pub := priv.PubKey()
-	info := newLocalInfo(name, pub, string(priv.Bytes()), algo)
-
-	kb.writeInfo(name, info)
-	return info
-}
-
-func (kb keyringKeybase) writeInfo(name string, info Info) {
-	// write the info by key
-	key := infoKey(name)
-	serializedInfo := marshalInfo(info)
-
-	err := kb.db.Set(keyring.Item{
-		Key:  string(key),
-		Data: serializedInfo,
-	})
+func (ks keystore) Sign(uid string, msg []byte) ([]byte, tmcrypto.PubKey, error) {
+	info, err := ks.Key(uid)
 	if err != nil {
-		panic(err)
+		return nil, nil, err
 	}
 
-	err = kb.db.Set(keyring.Item{
-		Key:  string(addrHexKey(info.GetAddress())),
-		Data: key,
-	})
-	if err != nil {
-		panic(err)
+	var priv tmcrypto.PrivKey
+
+	switch i := info.(type) {
+	case localInfo:
+		if i.PrivKeyArmor == "" {
+			return nil, nil, fmt.Errorf("private key not available")
+		}
+
+		priv, err = cryptoAmino.PrivKeyFromBytes([]byte(i.PrivKeyArmor))
+		if err != nil {
+			return nil, nil, err
+		}
+
+	case ledgerInfo:
+		return SignWithLedger(info, msg)
+
+	case offlineInfo, multiInfo:
+		return nil, info.GetPubKey(), errors.New("cannot sign with offline keys")
 	}
+
+	sig, err := priv.Sign(msg)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	return sig, priv.PubKey(), nil
+}
+
+func (ks keystore) SignByAddress(address sdk.Address, msg []byte) ([]byte, tmcrypto.PubKey, error) {
+	key, err := ks.KeyByAddress(address)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	return ks.Sign(key.GetName(), msg)
+}
+
+func (ks keystore) SaveLedgerKey(uid string, algo SigningAlgo, hrp string, account, index uint32) (Info, error) {
+	if !ks.options.supportedAlgosLedger.Contains(algo) {
+		return nil, ErrUnsupportedSigningAlgo
+	}
+
+	coinType := sdk.GetConfig().GetCoinType()
+	hdPath := hd.NewFundraiserParams(account, coinType, index)
+
+	priv, _, err := crypto.NewPrivKeyLedgerSecp256k1(*hdPath, hrp)
+	if err != nil {
+		return nil, err
+	}
+
+	return ks.writeLedgerKey(uid, priv.PubKey(), *hdPath, algo.Name())
+}
+
+func (ks keystore) writeLedgerKey(name string, pub tmcrypto.PubKey, path hd.BIP44Params, algo pubKeyType) (Info, error) {
+	info := newLedgerInfo(name, pub, path, algo)
+	err := ks.writeInfo(name, info)
+	if err != nil {
+		return nil, err
+	}
+
+	return info, nil
+}
+
+type altKrOptions struct {
+	supportedAlgos       AltSigningAlgoList
+	supportedAlgosLedger AltSigningAlgoList
+}
+
+func (ks keystore) SaveMultisig(uid string, pubkey tmcrypto.PubKey) (Info, error) {
+	return ks.writeMultisigKey(uid, pubkey)
+}
+
+func (ks keystore) SavePubKey(uid string, pubkey tmcrypto.PubKey, algo pubKeyType) (Info, error) {
+	return ks.writeOfflineKey(uid, pubkey, algo)
+}
+
+func (ks keystore) DeleteByAddress(address sdk.Address) error {
+	info, err := ks.KeyByAddress(address)
+	if err != nil {
+		return err
+	}
+
+	err = ks.Delete(info.GetName())
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func (ks keystore) Delete(uid string) error {
+	info, err := ks.Key(uid)
+	if err != nil {
+		return err
+	}
+
+	err = ks.db.Remove(addrHexKeyAsString(info.GetAddress()))
+	if err != nil {
+		return err
+	}
+
+	err = ks.db.Remove(string(infoKey(uid)))
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func (ks keystore) KeyByAddress(address sdk.Address) (Info, error) {
+	ik, err := ks.db.Get(addrHexKeyAsString(address))
+	if err != nil {
+		return nil, err
+	}
+
+	if len(ik.Data) == 0 {
+		return nil, fmt.Errorf("key with address %s not found", address)
+	}
+
+	bs, err := ks.db.Get(string(ik.Data))
+	if err != nil {
+		return nil, err
+	}
+
+	return unmarshalInfo(bs.Data)
+}
+
+func (ks keystore) List() ([]Info, error) {
+	var res []Info
+	keys, err := ks.db.Keys()
+	if err != nil {
+		return nil, err
+	}
+
+	sort.Strings(keys)
+
+	for _, key := range keys {
+		if strings.HasSuffix(key, infoSuffix) {
+			rawInfo, err := ks.db.Get(key)
+			if err != nil {
+				return nil, err
+			}
+
+			if len(rawInfo.Data) == 0 {
+				return nil, sdkerrors.Wrap(sdkerrors.ErrKeyNotFound, key)
+			}
+
+			info, err := unmarshalInfo(rawInfo.Data)
+			if err != nil {
+				return nil, err
+			}
+
+			res = append(res, info)
+		}
+	}
+
+	return res, nil
+}
+
+func (ks keystore) NewMnemonic(uid string, language Language, algo SigningAlgo) (Info, string, error) {
+	if language != English {
+		return nil, "", ErrUnsupportedLanguage
+	}
+
+	if !ks.isSupportedSigningAlgo(algo) {
+		return nil, "", ErrUnsupportedSigningAlgo
+	}
+
+	// Default number of words (24): This generates a mnemonic directly from the
+	// number of words by reading system entropy.
+	entropy, err := bip39.NewEntropy(defaultEntropySize)
+	if err != nil {
+		return nil, "", err
+	}
+
+	mnemonic, err := bip39.NewMnemonic(entropy)
+	if err != nil {
+		return nil, "", err
+	}
+
+	info, err := ks.NewAccount(uid, mnemonic, DefaultBIP39Passphrase, sdk.GetConfig().GetFullFundraiserPath(), algo)
+	if err != nil {
+		return nil, "", err
+	}
+
+	return info, mnemonic, err
+}
+
+func (ks keystore) NewAccount(uid string, mnemonic string, bip39Passphrase string, hdPath string, algo SigningAlgo) (Info, error) {
+	if !ks.isSupportedSigningAlgo(algo) {
+		return nil, ErrUnsupportedSigningAlgo
+	}
+
+	// create master key and derive first key for keyring
+	derivedPriv, err := algo.DeriveKey()(mnemonic, bip39Passphrase, hdPath)
+	if err != nil {
+		return nil, err
+	}
+
+	privKey := algo.PrivKeyGen()(derivedPriv)
+
+	return ks.writeLocalKey(uid, privKey, algo.Name())
+}
+
+func (ks keystore) isSupportedSigningAlgo(algo SigningAlgo) bool {
+	return ks.options.supportedAlgos.Contains(algo)
+}
+
+func (ks keystore) Key(uid string) (Info, error) {
+	key := infoKey(uid)
+
+	bs, err := ks.db.Get(string(key))
+	if err != nil {
+		return nil, err
+	}
+
+	if len(bs.Data) == 0 {
+		return nil, sdkerrors.Wrap(sdkerrors.ErrKeyNotFound, uid)
+	}
+
+	return unmarshalInfo(bs.Data)
 }
 
 func lkbToKeyringConfig(appName, dir string, buf io.Reader, test bool) keyring.Config {
@@ -648,6 +633,63 @@ func newRealPrompt(dir string, buf io.Reader) func(string) (string, error) {
 	}
 }
 
-func addrHexKey(address types.Address) []byte {
-	return []byte(addrHexKeyAsString(address))
+func (ks keystore) writeLocalKey(name string, priv tmcrypto.PrivKey, algo pubKeyType) (Info, error) {
+	// encrypt private key using keyring
+	pub := priv.PubKey()
+
+	info := newLocalInfo(name, pub, string(priv.Bytes()), algo)
+	err := ks.writeInfo(name, info)
+	if err != nil {
+		return nil, err
+	}
+
+	return info, nil
+}
+
+func (ks keystore) writeInfo(name string, info Info) error {
+	// write the info by key
+	key := infoKey(name)
+	serializedInfo := marshalInfo(info)
+
+	err := ks.db.Set(keyring.Item{
+		Key:  string(key),
+		Data: serializedInfo,
+	})
+	if err != nil {
+		return err
+	}
+
+	err = ks.db.Set(keyring.Item{
+		Key:  addrHexKeyAsString(info.GetAddress()),
+		Data: key,
+	})
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func (ks keystore) writeOfflineKey(name string, pub tmcrypto.PubKey, algo pubKeyType) (Info, error) {
+	info := newOfflineInfo(name, pub, algo)
+	err := ks.writeInfo(name, info)
+	if err != nil {
+		return nil, err
+	}
+
+	return info, nil
+}
+
+func (ks keystore) writeMultisigKey(name string, pub tmcrypto.PubKey) (Info, error) {
+	info := NewMultiInfo(name, pub)
+	err := ks.writeInfo(name, info)
+	if err != nil {
+		return nil, err
+	}
+
+	return info, nil
+}
+
+func addrHexKeyAsString(address sdk.Address) string {
+	return fmt.Sprintf("%s.%s", hex.EncodeToString(address.Bytes()), addressSuffix)
 }
